Removed BTIterator member from BehaviorTree
The BehaviorTree should not contain a reference to its end to be able to
assert whether two BTIterators are from the same BehaviorTree.

Now, the BTIterator contains a reference to the BehaviorTree it was created
from.

Also, BTIterator creation count was removed from BehaviorTree, and will now
depend on the user to make sure it is used responsibly.