Redesign layout
Removed the previously implemented layout for the new design's one.
This change affects both the CatLing code and the Test code.

This layout approximates the design. It has additions that have to be
added to the design. Such additions have been written down in paper, to
soon be added to the design.