Add simple ActionBehavior test
To be able to test the tree, the Commander is used. For starters, only
ActionBehavior is being tested, with a single iteration through the tree. This
is because there's no way to restart the tree yet, which means every time we
wanted to use it, it would have to be rebuilt from scratch.

To be able to do a simple test, a TWildcard Task was implemented, which will
simply call a given function.