Add multiple actions to sequence test
The implementation of the test was changed to make it a little more flexible
when creating more actions and adding them to the sequence. It's a complex
process, and complexity will grow with each new layer of behaviors. It has
become clear that an easier way of completing this process is required.