Remove ambiguity after implementing Slab's UnitType
For naming consistency, Slab's UnitType clashes with BWAPI's UnitType. At the
moment, names were disambiguated.

Another problem was detected when doing the corrections. The currently
implemented tasks, such as TGather or TBuild, require parameter passing on
construction, which removes the flexibility of the behavior tree and forces
the user to know the unit beforehand. This should be looked into, because it is
undesireable to have to build a tree for every situation, and because it's
impractical, in the sense that it forces to create very small trees, composed of
information retrieval, in order to 'feed' action oriented trees.