Add TSelectBuilder
TSelectBuilder is designed to obtain a worker from the Slab of free workers and
then move it into a Slab of builders.

However, there's a shortcoming. At the moment, several modules could retrieve
the same worker simultaneously. This could be overcome by adding a getAndDelete
function to the Slab class.