Change TBuild location parameter into function
Instead of giving the TBuild task a final position, we're going to try to give
it a function which can obtain the final position, when run. This is one of the
possible ways in which parameter passing could be implemented.

This is experimental, and may not be the final way in which parameter passing
is done.