Remove shared_ptr from slab initialization
Having a shared pointer was forcing deallocation on function return, which
automatically deallocated memory, thus leaving a pointer to a memory region in
the vector, and corrupting the data.

At the moment, header values are not deallocated, which could be a memory leak
if at some point we actually wanted to destroy slabs. Until then, because we're
never deallocating the memory on purpose, there's "no leak".