Begin resource allocation implementation
This is actually only a barebones of some of the implementation. In order to be
able to control resources more easily, we need a better implementation of the
Slabs. This is because the current slabs don't allow for random access, and to
be able to manage each entry, we have to use append functionality. This will
reorder the entries in the slab, and forces the module to keep track of where
its resources are.

This branch should be paused until a better implementation of slabs which solves
the problem is done.