Change BTIterator in order to make it directly accessible
The previous version of BTIterator only allowed it to be used by looping over it
in a for loop, which removed the ability to control when a behavior was called.
This forced the flow of the program to execute the whole tree in a single
Starcraft frame, which was undesired.