Reimplement TSelectBuilder
This reimplementation allows to select any builder from the worker pool that
isn't currently building or gathering gas. This was done in order to use it in
the plan that MacroPlanner will generate for building structures.

Also, it changes the API to return a boolean value like the rest of the Tasks.